server/requirements.txt
asgiref==3.7.2
boto3==1.28.37
botocore==1.31.52
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.5
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==21.0.0
gunicorn==23.0.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytest==7.4.3
pytest-django==4.5.2
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.2
sentry-sdk==2.13.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
